+++ lib/nested_select/deep_merger.rb
@@ -1,11 +1,12 @@
+# frozen_string_literal: true
+
 module NestedSelect
   module DeepMerger
-
     # {user_profile: [:zip_code]} + {user_profile: [:bio]} -> { user_profile: [:zip_code, :bio] }
     refine Hash do
       def deep_combine(other)
         merge!(other.except(*keys))
-        merge!(other.slice(*keys).map{ |key, value| [key, [self[key], value].flatten.deep_combine_elements]}.to_h )
+        merge!(other.slice(*keys).map { |key, value| [key, [self[key], value].flatten.deep_combine_elements] }.to_h)
       end
     end